+++ src/types/exports.ts
@@ -80,14 +80,17 @@ export interface EdgeContextProps extends CommonProps {
   apiKey?: string
   apiSecret?: Uint8Array
   appId?: string
-  authServer?: string
   infoServer?: string | string[]
+  loginServer: string | string[] // Do not include `/api` in the path
   syncServer?: string | string[]
   deviceDescription?: string
   hideKeys?: boolean
   logSettings?: Partial<EdgeLogSettings>
   plugins?: EdgeCorePluginsInit
   skipBlockHeight?: boolean
+
+  /** @deprecated Use `loginServer` instead */
+  authServer?: string
 }
 
 export interface EdgeFakeWorldProps extends CommonProps {